update help messages + main rework
I wasn't happy with the direction I was going with the commands module
so I decided it would be better to give the main module the
responsibility of parsing and mapping the arguments into the proper
handlers. Previously I was thinking the commands module would handle
pretty much everthing but I didn't like how I was forcing the command
handlers to deal with arguments. I prefer to pass all possble arguments
into the handlers instead. I think this is a better way to organize this
code.