use default Meta struct for init
Instead of hard coding a string for the initial Meta object, I'm
serializing the default object. I'm doing this for a couple of reasons.
First, I wanted to see if it would actually work and it did. Second,
I've been going back and forth with how I'm going to manage some things
related to the meta file so I'm going to try and leverage the type
system to help me out here....this second point is borerline bs but I'm
okay with it for now.