big rework
Two main things:
- argument handling moved to args.zig
- UX improvements for command output

This was a big change that I wanted to start on once I most of the main
command functionality done. This work cleans things up quite a bit and I
think it all reads much better now. It's not perfect, of course, but
it's moving in good direction.