"edit" and "delete" commands
Just went for it on this one because I started going down the path of
automatically deleting an edited file that was emptied, but then
realized it would be easier to just tell you that you emptied the file,
it's not cool, but it's up to you. By that point I had already got all
the "delete" command logic in, so I just moved it. At the same time I
modified the output of listing goals to very clearly mark the active
goal.