prepare-commit-hook + stdout
Alright now we're really cookin! `git init` creates a commit hook to
append the active goal details to commit messages. This also led to a
change that I needed to do anyways which is switching from debug print
to stdout. I'm still using debug print in the error cases but eventually
I'll switch to stderr...maybe. Also added a separate command for getting
the output into the commit message....called `commitmsg` which is a
slight modification of the `status` command.

------------------------------------------------------------------------

[ Goal #22 ] - init prepare-commit-hook

One of the main features of `goal` is linking goals to commits, so this
hook is crucial. I want this hook created in the `goal init` command.

------------------------------------------------------------------------